Make oj-bundle command
close #96
