+++ onlinejudge_verify/main.py
@@ -5,7 +5,6 @@ import math
 import os
 import pathlib
 import subprocess
-import sys
 from logging import DEBUG, basicConfig, getLogger
 from typing import *
 
@@ -27,10 +26,6 @@ def get_parser() -> argparse.ArgumentParser:
     subparser = subparsers.add_parser('run')
     subparser.add_argument('path', nargs='*', type=pathlib.Path)
     subparser.add_argument('-j', '--jobs', type=int, default=1)
-
-    subparser = subparsers.add_parser('bundle')
-    subparser.add_argument('path', type=pathlib.Path)
-    subparser.add_argument('-I', metavar='dir', type=pathlib.Path, dest='iquote', default=pathlib.Path.cwd(), help='add the directory dir to the list of directories to be searched for header files during preprocessing (default: ".")')
 
     subparser = subparsers.add_parser('docs')
 
@@ -150,12 +145,6 @@ def subcommand_docs() -> None:
         onlinejudge_verify.docs.main(html=False, force=True)
 
 
-def subcommand_bundle(path: pathlib.Path, *, iquote: pathlib.Path) -> None:
-    language = onlinejudge_verify.languages.get(path)
-    assert language is not None
-    sys.stdout.buffer.write(language.bundle(path, basedir=iquote))
-
-
 def main(args: Optional[List[str]] = None) -> None:
     basicConfig(level=DEBUG)
     parser = get_parser()
@@ -174,9 +163,6 @@ def main(args: Optional[List[str]] = None) -> None:
 
     elif parsed.subcommand == 'run':
         subcommand_run(paths=parsed.path, jobs=parsed.jobs)
-
-    elif parsed.subcommand == 'bundle':
-        subcommand_bundle(parsed.path, iquote=parsed.iquote)
 
     elif parsed.subcommand == 'docs':
         subcommand_docs()
